requirements.txt
fastapi
uvicorn[standard]
python-telegram-bot==20.8
google-genai
python-dotenv